+++ hh/src/main/java/com/hh/hh/salary/entity/PayrollDto.java
@@ -1,11 +1,11 @@
-package com.hh.salary;
+package com.hh.hh.salary.entity;
 
 import java.util.Date;
 
 import lombok.Data;
 
 @Data
-public class SalaryDto {
+public class PayrollDto {
 
 	private int empNo;
 	private String basicSalary;
@@ -24,5 +24,7 @@ public class SalaryDto {
 	private String fourMajorInsurances;
 	private String tax;
 	private String netSalary;
+	private String payrollAccount;
+	private Date payrollDate;
 
 }
