+++ hh/src/main/webapp/WEB-INF/views/salary/payroll.jsp
@@ -1,4 +1,4 @@
-<%@page import="com.hh.salary.PayrollDto"%>
+<%@page import="com.hh.hh.salary.entity.PayrollDto"%>
 <%@page import="java.util.List"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
